refactor: move CreateFlow
